backend/requirements.txt
Flask==3.0.3
Flask-CORS==4.0.1
python-dotenv==1.0.1
supabase==2.5.0
pytesseract==0.3.10
Pillow>=10.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
joblib>=1.3.0
pandas>=2.0.0
pdf2image>=1.16.0
PyPDF2>=3.0.0
opencv-python>=4.8.0
werkzeug>=3.0.0
gunicorn>=22.0.0
